+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "1D bush fire".
 - Tests for "Rectangle partition".
 - Tests for "Robot show".
+- Tests for "Rock paper scissors lizard Spock".
 
 ## [1.4.0] - 2022-03-03
 ### Added